refactor(patient_relationships): rename module patient_admin -> patient_relationships
"patient_admin" was too generic for a module that only holds
patient-to-patient links. Nothing has shipped, so this is a clean
rename with no compat aliases:

- module/entry-point/branch name, docs dir, tests dir
- table patient_admin_relationship -> patient_relationships (+ index
  and constraint names); exemption-status table in the migration pair
  renamed to match the new prefix
- revisions padm_0001/padm_0002 -> prel_0001/prel_0002
- API prefix /api/v1/patient_relationships (loader mounts by module name)
- permissions collapsed to patient_relationships.read/.write — bare
  read/write from get_permissions(), the patients/recalls convention,
  which kills the old patient_admin.relationships.* double-noun
- frontend: PatientRelationshipsCard.vue, usePatientRelationships.ts,
  patientRelationships i18n keys + permissions.ts entry, slot id
- catalogs regenerated

Existing dev installs of the old name must be uninstalled or manually
dropped (alembic_version padm_* rows, patient_admin_relationship table,
core_module row) — no migration path, per the CHANGELOG note.

Verified after rename: ruff clean; module tests + manifest validator
10 passed; alembic_roundtrip suite 9 passed; eslint clean; nuxt
typecheck (host + 23 layers) clean; generate_catalogs --check,
check_docs_coverage --strict and check_docs_layout green.
