+++ backend/alembic.ini
@@ -8,7 +8,7 @@ version_path_separator = os
 # full graph before ``env.py`` loads. ``env.py`` still merges in any
 # additional branches discovered at runtime (e.g. community modules
 # installed from PyPI) via ``discover_version_locations``.
-version_locations = alembic/versions:app/modules/patients/migrations/versions:app/modules/patients_clinical/migrations/versions:app/modules/agenda/migrations/versions:app/modules/patient_timeline/migrations/versions:app/modules/catalog/migrations/versions:app/modules/odontogram/migrations/versions:app/modules/treatment_plan/migrations/versions:app/modules/budget/migrations/versions:app/modules/billing/migrations/versions:app/modules/media/migrations/versions:app/modules/notifications/migrations/versions:app/modules/payments/migrations/versions:app/modules/schedules/migrations/versions:app/modules/verifactu/migrations/versions:app/modules/clinical_notes/migrations/versions:app/modules/recalls/migrations/versions:app/modules/migration_import/migrations/versions:app/modules/periodontogram/migrations/versions:app/modules/copilot/migrations/versions:app/modules/accounting_export/migrations/versions:app/modules/whatsapp_kapso/migrations/versions:app/modules/patient_admin/migrations/versions
+version_locations = alembic/versions:app/modules/patients/migrations/versions:app/modules/patients_clinical/migrations/versions:app/modules/agenda/migrations/versions:app/modules/patient_timeline/migrations/versions:app/modules/catalog/migrations/versions:app/modules/odontogram/migrations/versions:app/modules/treatment_plan/migrations/versions:app/modules/budget/migrations/versions:app/modules/billing/migrations/versions:app/modules/media/migrations/versions:app/modules/notifications/migrations/versions:app/modules/payments/migrations/versions:app/modules/schedules/migrations/versions:app/modules/verifactu/migrations/versions:app/modules/clinical_notes/migrations/versions:app/modules/recalls/migrations/versions:app/modules/migration_import/migrations/versions:app/modules/periodontogram/migrations/versions:app/modules/copilot/migrations/versions:app/modules/accounting_export/migrations/versions:app/modules/whatsapp_kapso/migrations/versions:app/modules/patient_relationships/migrations/versions
 
 [post_write_hooks]
