--- backend/app/modules/patient_admin/frontend/i18n/locales/en.json
@@ -1,22 +1,0 @@
-{
-  "patientAdmin": {
-    "title": "Family",
-    "emptyHint": "No family links on file.",
-    "relationships": {
-      "title": "Family relationships",
-      "searchPatient": "Search patient to link…",
-      "addSuccess": "Relationship added",
-      "addError": "Could not add relationship",
-      "removeError": "Could not remove relationship",
-      "types": {
-        "parent": "Parent",
-        "child": "Child",
-        "spouse": "Spouse",
-        "sibling": "Sibling",
-        "guardian": "Guardian",
-        "ward": "Ward",
-        "other": "Other"
-      }
-    }
-  }
-}
