--- backend/app/modules/patient_admin/frontend/i18n/locales/ta.json
@@ -1,22 +1,0 @@
-{
-  "patientAdmin": {
-    "title": "குடும்பம்",
-    "emptyHint": "குடும்ப இணைப்புகள் பதிவு செய்யப்படவில்லை.",
-    "relationships": {
-      "title": "குடும்ப உறவுகள்",
-      "searchPatient": "இணைக்க நோயாளியைத் தேடுங்கள்…",
-      "addSuccess": "உறவு சேர்க்கப்பட்டது",
-      "addError": "உறவைச் சேர்க்க முடியவில்லை",
-      "removeError": "உறவை அகற்ற முடியவில்லை",
-      "types": {
-        "parent": "பெற்றோர்",
-        "child": "குழந்தை",
-        "spouse": "துணைவர்",
-        "sibling": "உடன்பிறப்பு",
-        "guardian": "பாதுகாவலர்",
-        "ward": "பராமரிப்பில் உள்ளவர்",
-        "other": "மற்றவை"
-      }
-    }
-  }
-}
