--- backend/app/modules/patient_admin/frontend/nuxt.config.ts
@@ -1,13 +1,0 @@
-// Nuxt layer for the `patient_admin` module.
-export default defineNuxtConfig({
-  i18n: {
-    locales: [
-      { code: 'en', file: 'en.json' },
-      { code: 'es', file: 'es.json' },
-      { code: 'fr', file: 'fr.json' },
-      { code: 'pt', file: 'pt.json' },
-      { code: 'ta', file: 'ta.json' }
-    ],
-    langDir: 'locales'
-  }
-})
