+++ backend/app/modules/patient_relationships/frontend/nuxt.config.ts
@@ -1,0 +1,13 @@
+// Nuxt layer for the `patient_relationships` module.
+export default defineNuxtConfig({
+  i18n: {
+    locales: [
+      { code: 'en', file: 'en.json' },
+      { code: 'es', file: 'es.json' },
+      { code: 'fr', file: 'fr.json' },
+      { code: 'pt', file: 'pt.json' },
+      { code: 'ta', file: 'ta.json' }
+    ],
+    langDir: 'locales'
+  }
+})
