+++ backend/pyproject.toml
@@ -76,7 +76,7 @@ recalls = "app.modules.recalls:RecallsModule"
 verifactu = "app.modules.verifactu:VerifactuModule"
 migration_import = "app.modules.migration_import:MigrationImportModule"
 copilot = "app.modules.copilot:CopilotModule"
-patient_admin = "app.modules.patient_admin:PatientAdminModule"
+patient_relationships = "app.modules.patient_relationships:PatientRelationshipsModule"
 
 [tool.setuptools.packages.find]
 where = ["."]
