+++ docs/modules-catalog.md
@@ -21,7 +21,7 @@ Maintained by `backend/scripts/generate_catalogs.py`. CI fails if a manifest cha
 | `migration_import` | 0.1.0 | official | patients, patients_clinical, clinical_notes, agenda, schedules, recalls, catalog, budget, odontogram, treatment_plan, billing, payments, media | manual | yes | 4 | 5 | 0 | yes |
 | `notifications` | 0.1.0 | official | patients, agenda, budget, billing, catalog | auto | no | 8 | 7 | 6 | yes |
 | `odontogram` | 0.3.0 | official | patients, catalog | auto | no | 4 | 7 | 0 | yes |
-| `patient_admin` | 0.2.0 | community | patients | manual | yes | 2 | 0 | 0 | yes |
+| `patient_relationships` | 0.2.0 | community | patients | manual | yes | 2 | 0 | 0 | yes |
 | `patient_timeline` | 0.1.0 | official | patients | auto | no | 1 | 0 | 35 | yes |
 | `patients` | 0.1.0 | official | — | auto | no | 2 | 3 | 0 | yes |
 | `patients_clinical` | 0.1.0 | official | patients | auto | no | 4 | 1 | 0 | yes |
@@ -319,7 +319,7 @@ Dental charting, tooth state, clinical treatments.
 - **Events consumed:** —
 - **Module CLAUDE.md:** [`backend/app/modules/odontogram/CLAUDE.md`](../backend/app/modules/odontogram/CLAUDE.md)
 
-### `patient_admin` — v0.2.0
+### `patient_relationships` — v0.2.0
 
 Patient family relationships (Lien de Parentée).
 
@@ -330,11 +330,11 @@ Patient family relationships (Lien de Parentée).
 - **Depends:** `patients`
 - **Frontend layer:** `frontend`
 - **Permissions:**
-  - `patient_admin.relationships.read`
-  - `patient_admin.relationships.write`
+  - `patient_relationships.read`
+  - `patient_relationships.write`
 - **Events emitted:** —
 - **Events consumed:** —
-- **Module CLAUDE.md:** [`backend/app/modules/patient_admin/CLAUDE.md`](../backend/app/modules/patient_admin/CLAUDE.md)
+- **Module CLAUDE.md:** [`backend/app/modules/patient_relationships/CLAUDE.md`](../backend/app/modules/patient_relationships/CLAUDE.md)
 
 ### `patient_timeline` — v0.1.0
